requirements.txt
Django==5.2.1
python-decouple